refactor(bulk-upload): Break down routes into individual files
